+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## 2.0.0
+ * STORM-1553: port event.clj to java
+ * STORM-1262: port backtype.storm.command.dev-zookeeper to java.
+ * STORM-1243: port backtype.storm.command.healthcheck to java.
+ * STORM-1246: port backtype.storm.local-state to java.
+ * STORM-1516: Fixed issue in writing pids with distributed cluster mode.
  * STORM-1253: port backtype.storm.timer to java
  * STORM-1258: port thrift.clj to Thrift.java
  * STORM-1336: Evalute/Port JStorm cgroup support and implement cgroup support for RAS